pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "online_research_engine"
version = "2024.9.6"
authors = [
  { name="Wilson ZheLin" },
  { name="James Palmer" },
]
description = "A package that enables LLMs to access online content in real-time when answering a query."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "beautifulsoup4==4.12.2",
  "chromadb==0.4.18",
  "langchain==0.0.340",
  "lxml==4.9.3",
  "openai==1.3.4",
  "PyYAML==6.0.1",
  "Requests==2.31.0",
  "tiktoken==0.7.0",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/jamespalmer2000/GPT-4-Web-Browsing"
Repository = "https://github.com/jamespalmer2000/GPT-4-Web-Browsing"
Issues = "https://github.com/jamespalmer2000/GPT-4-Web-Browsing/issues"